test: relocate notification test
